File transfer implementation testing phase 1 complete

+++ app/src/main/java/com/aj/sendall/db/util/DBUtil.java
@@ -58,7 +58,7 @@ public class DBUtil {
     }
 
     @SuppressWarnings("unchecked")
-    public List<PersonalInteractionDTO> getAllPersonalInteractionDTO(Integer connId){
+    public List<PersonalInteractionDTO> getAllPersonalInteractionDTO(long connId){
         List<PersonalInteraction> pis = daoSession.getPersonalInteractionDao()
                 .queryBuilder()
                 .where(PersonalInteractionDao.Properties.ConnectionId.eq(connId))
